Make generators nested in tasks
