+++ lab-01/solution/src/by/busskov/quizer/Operation.java
