+++ lab-01/solution/src/by/busskov/quizer/Task.java
@@ -1,6 +1,10 @@
 package by.busskov.quizer;
 
 public interface Task {
+    public interface Generator {
+        Task generate();
+    }
+
     String getText();
 
     Result validate(String answer);
