--- lab-01/solution/src/by/busskov/quizer/TaskGenerator.java
@@ -1,5 +1,0 @@
-package by.busskov.quizer;
-
-public interface TaskGenerator {
-    Task generate();
-}
