+++ lab-01/solution/src/by/busskov/quizer/task_generators/GroupTaskGenerator.java
@@ -1,27 +1,26 @@
 package by.busskov.quizer.task_generators;
 
 import by.busskov.quizer.Task;
-import by.busskov.quizer.TaskGenerator;
 
 import java.util.*;
 
-public class GroupTaskGenerator implements TaskGenerator {
-    public GroupTaskGenerator(TaskGenerator... generators) {
+public class GroupTaskGenerator implements Task.Generator {
+    public GroupTaskGenerator(Task.Generator... generators) {
         this.generators = Arrays.copyOf(generators, generators.length);
     }
 
-    public GroupTaskGenerator(Collection<? extends TaskGenerator> generators) {
+    public GroupTaskGenerator(Collection<? extends Task.Generator> generators) {
         Object[] array = generators.toArray();
-        this.generators = new TaskGenerator[array.length];
+        this.generators = new Task.Generator[array.length];
         for (int i = 0; i < array.length; ++i) {
-            this.generators[i] = (TaskGenerator) array[i];
+            this.generators[i] = (Task.Generator) array[i];
         }
     }
 
     @Override
     public Task generate() {
         Random random = new Random();
-        ArrayList<TaskGenerator> generatorsCopy = new ArrayList<>(List.of(generators));
+        ArrayList<Task.Generator> generatorsCopy = new ArrayList<>(List.of(generators));
         while (!generatorsCopy.isEmpty()) {
             int index = random.nextInt(generatorsCopy.size());
             try {
@@ -33,5 +32,5 @@ public class GroupTaskGenerator implements TaskGenerator {
         throw new IllegalStateException("All generators threw exceptions");
     }
 
-    private final TaskGenerator[] generators;
+    private final Task.Generator[] generators;
 }
