--- lab-01/solution/src/by/busskov/quizer/task_generators/math_task_generators/MathTaskGenerator.java
@@ -1,11 +1,0 @@
-package by.busskov.quizer.task_generators.math_task_generators;
-
-import by.busskov.quizer.TaskGenerator;
-
-public interface MathTaskGenerator extends TaskGenerator {
-    int getMinNumber();
-    int getMaxNumber();
-    default int getDiffNumber() {
-        return getMaxNumber() - getMinNumber();
-    }
-}
